+++ airflow/gcp/hooks/spanner.py
@@ -78,7 +78,8 @@ class SpannerHook(CloudBaseHook):
         :return: Spanner instance
         :rtype: google.cloud.spanner_v1.instance.Instance
         """
-        assert project_id is not None
+        if not project_id:
+            raise ValueError("The project_id should be set")
         instance = self._get_client(project_id=project_id).instance(instance_id=instance_id)
         if not instance.exists():
             return None
@@ -156,7 +157,8 @@ class SpannerHook(CloudBaseHook):
         :type project_id: str
         :return: None
         """
-        assert project_id is not None
+        if not project_id:
+            raise ValueError("The project_id should be set")
         self._apply_to_instance(project_id, instance_id, configuration_name,
                                 node_count, display_name, lambda x: x.create())
 
@@ -190,7 +192,8 @@ class SpannerHook(CloudBaseHook):
         :type project_id: str
         :return: None
         """
-        assert project_id is not None
+        if not project_id:
+            raise ValueError("The project_id should be set")
         self._apply_to_instance(project_id, instance_id, configuration_name,
                                 node_count, display_name, lambda x: x.update())
 
@@ -206,7 +209,8 @@ class SpannerHook(CloudBaseHook):
         :type project_id: str
         :return: None
         """
-        assert project_id is not None
+        if not project_id:
+            raise ValueError("The project_id should be set")
         instance = self._get_client(project_id=project_id).instance(instance_id)
         try:
             instance.delete()
@@ -236,7 +240,8 @@ class SpannerHook(CloudBaseHook):
         :return: Database object or None if database does not exist
         :rtype: google.cloud.spanner_v1.database.Database or None
         """
-        assert project_id is not None
+        if not project_id:
+            raise ValueError("The project_id should be set")
 
         instance = self._get_client(project_id=project_id).instance(
             instance_id=instance_id)
@@ -271,7 +276,8 @@ class SpannerHook(CloudBaseHook):
             database. If set to None or missing, the default project_id from the GCP connection is used.
         :return: None
         """
-        assert project_id is not None
+        if not project_id:
+            raise ValueError("The project_id should be set")
 
         instance = self._get_client(project_id=project_id).instance(
             instance_id=instance_id)
@@ -316,7 +322,8 @@ class SpannerHook(CloudBaseHook):
         :type operation_id: str
         :return: None
         """
-        assert project_id is not None
+        if not project_id:
+            raise ValueError("The project_id should be set")
 
         instance = self._get_client(project_id=project_id).instance(
             instance_id=instance_id)
@@ -355,7 +362,8 @@ class SpannerHook(CloudBaseHook):
         :return: True if everything succeeded
         :rtype: bool
         """
-        assert project_id is not None
+        if not project_id:
+            raise ValueError("The project_id should be set")
 
         instance = self._get_client(project_id=project_id).\
             instance(instance_id=instance_id)
@@ -398,7 +406,8 @@ class SpannerHook(CloudBaseHook):
             database. If set to None or missing, the default project_id from the GCP connection is used.
         :type project_id: str
         """
-        assert project_id is not None
+        if not project_id:
+            raise ValueError("The project_id should be set")
 
         self._get_client(project_id=project_id).instance(instance_id=instance_id).\
             database(database_id=database_id).run_in_transaction(
